intro_home, order, roulette, calibration
