+++ FoodInEye_app/app/src/main/AndroidManifest.xml
@@ -24,13 +24,13 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".Intro4Activity"
+            android:name=".activity.Intro4Activity"
             android:exported="false" />
         <activity
-            android:name=".Intro3Activity"
+            android:name=".activity.Intro3Activity"
             android:exported="false" />
         <activity
-            android:name=".Intro2Activity"
+            android:name=".activity.Intro2Activity"
             android:exported="false" />
         <activity
             android:name=".activity.IntroActivity"
