--- FoodInEye_app/app/src/main/java/com/example/foodineye_app/Intro4Activity.java
@@ -1,14 +1,0 @@
-package com.example.foodineye_app;
-
-import android.os.Bundle;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-public class Intro4Activity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_intro4);
-    }
-}
