+++ FoodInEye_app/app/src/main/java/com/example/foodineye_app/activity/HomeActivity.java
@@ -153,6 +153,8 @@ public class HomeActivity extends AppCompatActivity {
 
                 Intent introIntent = new Intent(getApplicationContext(), IntroActivity.class);
                 startActivity(introIntent);
+                overridePendingTransition(R.anim.fade_in, R.anim.fade_out);
+                finish();
 
             }
         });
