+++ FoodInEye_app/app/src/main/java/com/example/foodineye_app/activity/Intro3Activity.java
@@ -11,7 +11,7 @@ import androidx.appcompat.widget.Toolbar;
 
 import com.example.foodineye_app.R;
 
-public class IntroActivity extends AppCompatActivity {
+public class Intro3Activity extends AppCompatActivity {
 
     Toolbar toolbar;
     Button nextBtn;
@@ -19,24 +19,23 @@ public class IntroActivity extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_intro);
+        setContentView(R.layout.activity_intro3);
 
         //toolbar
-        toolbar = (Toolbar) findViewById(R.id.intro_toolbar);
+        toolbar = (androidx.appcompat.widget.Toolbar) findViewById(R.id.intro3_toolbar);
         setToolBar(toolbar);
 
         //nettBtn
-        nextBtn = (Button) findViewById(R.id.intro_next_btn);
+        nextBtn = (Button) findViewById(R.id.intro3_next_btn);
         nextBtn.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                Intent introIntent = new Intent(getApplicationContext(), Intro2Activity.class);
+                Intent introIntent = new Intent(getApplicationContext(), Intro4Activity.class);
                 startActivity(introIntent);
                 overridePendingTransition(R.anim.fade_in, R.anim.fade_out);
                 finish();
             }
         });
-
     }
 
     //toolbar
@@ -46,7 +45,7 @@ public class IntroActivity extends AppCompatActivity {
         setSupportActionBar(toolbar);
 
         getSupportActionBar().setTitle(""); // 툴바의 타이틀을 직접 설정
-        ImageView backBtn = (ImageView) findViewById(R.id.intro_back);
+        ImageView backBtn = (ImageView) findViewById(R.id.intro3_back);
         backBtn.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
@@ -55,7 +54,7 @@ public class IntroActivity extends AppCompatActivity {
             }
         });
 
-        ImageView homeBtn = (ImageView) findViewById(R.id.intro_home);
+        ImageView homeBtn = (ImageView) findViewById(R.id.intro3_home);
         homeBtn.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
@@ -63,8 +62,6 @@ public class IntroActivity extends AppCompatActivity {
                 Intent loginIntent = new Intent(getApplicationContext(), HomeActivity.class);
                 startActivity(loginIntent);
                 finish();
-
-
             }
         });
