+++ FoodInEye_app/app/src/main/res/layout/activity_intro2.xml
@@ -7,7 +7,7 @@
 
     <!-- 커스텀 툴바 추가 -->
     <androidx.appcompat.widget.Toolbar
-        android:id="@+id/camera_toolbar"
+        android:id="@+id/intro2_toolbar"
         android:layout_width="match_parent"
         android:layout_height="?attr/actionBarSize"
         android:background="?attr/colorOnPrimary"
@@ -15,7 +15,7 @@
 
         <!-- 사용자 정의 뒤로 가기 버튼 추가 -->
         <ImageView
-            android:id="@+id/camera_back"
+            android:id="@+id/intro2_back"
             android:layout_width="40dp"
             android:layout_height="24dp"
             android:layout_gravity="start"
@@ -23,7 +23,7 @@
             android:contentDescription="뒤로 가기" />
 
         <TextView
-            android:id="@+id/camera_title"
+            android:id="@+id/intro2_title"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="한 눈에 학식"
@@ -32,7 +32,7 @@
             android:textColor="@color/black"/>
 
         <ImageView
-            android:id="@+id/camera_home"
+            android:id="@+id/intro2_home"
             android:layout_width="24dp"
             android:layout_height="24dp"
             android:src="@drawable/home"
@@ -41,61 +41,51 @@
 
     </androidx.appcompat.widget.Toolbar>
 
+    <ImageView
+        android:layout_width="match_parent"
+        android:layout_height="20dp"
+        android:src="@drawable/intro_2"
+        android:layout_gravity="center"
+        android:layout_marginTop="15dp"/>
+
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="시선 최적화"
+        android:textSize="30dp"
+        android:layout_gravity="center"
+        android:layout_marginTop="20dp"/>
+
     <LinearLayout
         android:layout_width="match_parent"
-        android:layout_height="180dp"
-        android:orientation="vertical">
-
-        <ImageView
-            android:layout_width="match_parent"
-            android:layout_height="20dp"
-            android:src="@drawable/intro_2"
-            android:layout_gravity="center"
-            android:layout_marginTop="15dp"/>
+        android:layout_height="530dp"
+        android:orientation="vertical"
+        android:background="@drawable/rec_str10"
+        android:layout_margin="10dp"
+        android:padding="20dp">
 
         <TextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="시선 최적화"
-            android:textSize="30dp"
-            android:layout_gravity="center"
-            android:layout_marginTop="30dp"/>
+            android:text=" 사용자가 주문하기 전 시선 추적의 정확도를 높이기 위한 \n조정 작업입니다.\n 사용자의 시선에 맞게 시선 추적 기술과 관련된 기능을 이용할 수 있어요"
+            android:textSize="13dp"
+            android:lineSpacingExtra="3dp"
+            android:layout_gravity="center"/>
 
-        <TextView
-            android:layout_width="wrap_content"
+        <ImageView
+            android:layout_width="250dp"
             android:layout_height="wrap_content"
-            android:text="나의 시선으로 앱을 최적화 할 수 있어요."
-            android:textSize="16dp"
             android:layout_gravity="center"
-            android:layout_marginTop="30dp"
-            android:layout_marginBottom="30dp"/>
-
-    </LinearLayout>
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="400dp"
-        android:orientation="horizontal"
-        android:layout_margin="10dp">
-
-        <ImageView
-            android:layout_width="match_parent"
-            android:layout_height="400dp"
-            android:src="@drawable/next"
+            android:src="@drawable/intro_calibration"
             android:layout_weight="1"
-            android:layout_marginRight="2dp"/>
+            android:layout_marginTop="15dp"
+            android:layout_marginLeft="16dp"/>
 
-        <ImageView
-            android:layout_width="match_parent"
-            android:layout_height="400dp"
-            android:src="@drawable/next"
-            android:layout_weight="1"
-            android:layout_marginRight="2dp"/>
 
     </LinearLayout>
 
     <Button
-        android:id="@+id/next_btn"
+        android:id="@+id/intro2_next_btn"
         android:layout_width="140dp"
         android:layout_height="60dp"
         android:text="다음"
@@ -104,7 +94,7 @@
         android:background="@drawable/sign_up"
         android:backgroundTint="@color/blue"
         android:layout_gravity="center"
-        android:layout_marginTop="25dp"
+        android:layout_marginTop="20dp"
         android:layout_marginBottom="30dp"/>
 
 </LinearLayout>
