+++ FoodInEye_app/app/src/main/res/layout/activity_intro4.xml
@@ -8,7 +8,7 @@
 
     <!-- 커스텀 툴바 추가 -->
     <androidx.appcompat.widget.Toolbar
-        android:id="@+id/camera_toolbar"
+        android:id="@+id/intro4_toolbar"
         android:layout_width="match_parent"
         android:layout_height="?attr/actionBarSize"
         android:background="?attr/colorOnPrimary"
@@ -16,7 +16,7 @@
 
         <!-- 사용자 정의 뒤로 가기 버튼 추가 -->
         <ImageView
-            android:id="@+id/camera_back"
+            android:id="@+id/intro4_back"
             android:layout_width="40dp"
             android:layout_height="24dp"
             android:layout_gravity="start"
@@ -24,7 +24,7 @@
             android:contentDescription="뒤로 가기" />
 
         <TextView
-            android:id="@+id/camera_title"
+            android:id="@+id/intro4_title"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="한 눈에 학식"
@@ -33,7 +33,7 @@
             android:textColor="@color/black"/>
 
         <ImageView
-            android:id="@+id/camera_home"
+            android:id="@+id/intro4_home"
             android:layout_width="24dp"
             android:layout_height="24dp"
             android:src="@drawable/home"
@@ -42,10 +42,6 @@
 
     </androidx.appcompat.widget.Toolbar>
 
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="180dp"
-            android:orientation="vertical">
 
         <ImageView
             android:layout_width="match_parent"
@@ -62,40 +58,49 @@
             android:layout_gravity="center"
             android:layout_marginTop="30dp"/>
 
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="530dp"
+        android:orientation="vertical"
+        android:background="@drawable/rec_str10"
+        android:layout_margin="10dp"
+        android:padding="20dp">
+
         <TextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="나의 시선 속 메뉴들을 룰렛으로 선택하여 \n주문해보세요."
-            android:textSize="16dp"
+            android:text="나의 시선 속 메뉴들을 룰렛으로 선택하여 주문해보세요."
+            android:textSize="13dp"
             android:layout_gravity="center"
-            android:layout_marginTop="30dp"/>
+            android:layout_marginTop="10dp"/>
 
-    </LinearLayout>
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="400dp"
-        android:orientation="horizontal"
-        android:layout_margin="10dp">
-
-        <ImageView
+        <LinearLayout
             android:layout_width="match_parent"
             android:layout_height="400dp"
-            android:src="@drawable/next"
-            android:layout_weight="1"
-            android:layout_marginRight="2dp"/>
+            android:orientation="horizontal"
+            android:layout_marginTop="10dp">
 
-        <ImageView
-            android:layout_width="match_parent"
-            android:layout_height="400dp"
-            android:src="@drawable/next"
-            android:layout_weight="1"
-            android:layout_marginRight="2dp"/>
+            <ImageView
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:src="@drawable/intro_roulette1"
+                android:layout_weight="1"
+                android:layout_marginTop="15dp"/>
+
+            <ImageView
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:src="@drawable/intro_roulette"
+                android:layout_weight="1"
+                android:layout_marginTop="15dp"/>
+
+        </LinearLayout>
+
 
     </LinearLayout>
 
     <Button
-        android:id="@+id/next_btn"
+        android:id="@+id/intro4_next_btn"
         android:layout_width="140dp"
         android:layout_height="60dp"
         android:text="다음"
@@ -104,9 +109,8 @@
         android:background="@drawable/sign_up"
         android:backgroundTint="@color/blue"
         android:layout_gravity="center"
-        android:layout_marginTop="25dp"
+        android:layout_marginTop="20dp"
         android:layout_marginBottom="30dp"/>
-
 
 
 </LinearLayout>
